# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger de PurrfectMatch",
        "title": "GO-challenge-PurrfectMatch",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/associations": {
            "get": {
                "description": "Retrieve all associations from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Get all associations",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all associations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Association"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new association with the input payload and a PDF file",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Create a new association",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "AddressRue",
                        "name": "addressRue",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CP",
                        "name": "cp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ville",
                        "name": "ville",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Phone",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OwnerID",
                        "name": "ownerId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of members IDs",
                        "name": "members",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "PDF file",
                        "name": "kbisFile",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created association",
                        "schema": {
                            "$ref": "#/definitions/models.Association"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/associations/{id}": {
            "get": {
                "description": "Retrieve an association by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Get association by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Association ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved association",
                        "schema": {
                            "$ref": "#/definitions/models.Association"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid association ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found: Association not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update all fields of an association with the given ID",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Update an association",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Association ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "AddressRue",
                        "name": "addressRue",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "CP",
                        "name": "cp",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Ville",
                        "name": "ville",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Phone",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "PDF file",
                        "name": "kbisFile",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of members IDs",
                        "name": "members",
                        "in": "formData"
                    },
                    {
                        "description": "Association payload",
                        "name": "association",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Association"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated association",
                        "schema": {
                            "$ref": "#/definitions/models.Association"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid association ID or Invalid content type for kbisFile, expected application/pdf",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an association by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Delete an association",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Association ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted association"
                    },
                    "400": {
                        "description": "Bad Request: Invalid association ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found: Association not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/associations/{id}/verify": {
            "put": {
                "description": "Update the verify status of an association with the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Update an association's verify status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Association ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Verify status",
                        "name": "verified",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated association",
                        "schema": {
                            "$ref": "#/definitions/models.Association"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Missing association ID, Invalid association ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/users/{userId}/associations": {
            "get": {
                "description": "Retrieve all associations for a specific user where the user is either the owner or a member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "associations"
                ],
                "summary": "Get associations by user ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved associations for user",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Association"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid user ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Association": {
            "type": "object",
            "properties": {
                "addressRue": {
                    "type": "string"
                },
                "cp": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kbisFile": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                },
                "ville": {
                    "type": "string"
                }
            }
        },
...
    }
}